(1) Problems with marks and destroying peers (crash) solved. (2) Problems with wrap mode "codepoint" (wrong data/offsets) solved. (3) More assertions for wrap mode "codepoint". (4) Function MakeStyle() a bit overworked (array instead of multiple variables). (5) Function TkTextFreeMarks() fixed, because MarkDeleteProc() is deleting hash entries, and this should not happen while enumeration of hash table is in progress. (6) 'hyphens' in TkText remaned to 'useHyphenSupport' to avoid confusions with 'hyphenate'. (7) Wrong check in CheckLineMetricConsistency() removed.

+++ generic/tkTextTag.c
@@ -2423,7 +2423,7 @@ TagSortProc(
     const void *first,
     const void *second)		/* Elements to be compared. */
 {
-    return (*(TkTextTag **) first)->priority - (*(TkTextTag **) second)->priority;
+    return (int) (*(TkTextTag **) first)->priority - (int) (*(TkTextTag **) second)->priority;
 }
 
 void
@@ -3364,7 +3364,7 @@ TagBindEvent(
 	TkTextTag *tagPtr;
 
 	/*
-	 * Take into account that same tags have been gone in the meanwhile.
+	 * Take into account that some tags have been gone in the meanwhile.
 	 */
 
 	if (i >= maxTags) {
